Downgrade kernel only for mac
68c6383 moved the kernel changes from an aarch64-only block to a
block which is run when the macOS bundle is enabled. However, the
kernel downgrade is done before any bundle is generated, so the kernel
will be downgraded for all our bundles, linux, macos, windows.

This PR fixes this and use https://github.com/crc-org/snc/pull/637 logic
to only downgrade the kernel for macOS.

+++ createdisk.sh
@@ -138,33 +138,6 @@ fi
 
 cleanup_vm_image ${VM_NAME} ${VM_IP}
 
-# Only used for macOS bundle generation
-if [ "${SNC_GENERATE_MACOS_BUNDLE}" != "0" ]; then
-    if [ ${BUNDLE_TYPE} != "okd" ]; then
-        # workaround https://github.com/crc-org/vfkit/issues/11 on macOS 12
-        downgrade_rhel9_kernel ${VM_IP}
-        cleanup_vm_image ${VM_NAME} ${VM_IP}
-    fi 
-
-    # Get the rhcos ostree Hash ID
-    ostree_hash=$(${SSH} core@${VM_IP} -- "cat /proc/cmdline | grep -oP \"(?<=${BASE_OS}-).*(?=/vmlinuz)\"")
-
-    # Get the rhcos kernel release
-    kernel_release=$(${SSH} core@${VM_IP} -- 'uname -r')
-
-    # Get the kernel command line arguments
-    kernel_cmd_line=$(${SSH} core@${VM_IP} -- 'cat /proc/cmdline')
-
-    # Get the vmlinux/initramfs to /tmp/kernel and change permission for initramfs
-    ${SSH} core@${VM_IP} -- "mkdir /tmp/kernel && sudo cp -r /boot/ostree/${BASE_OS}-${ostree_hash}/*${kernel_release}* /tmp/kernel && sudo chmod 644 /tmp/kernel/initramfs*"
-
-    # SCP the vmlinuz/initramfs from VM to Host in provided folder.
-    ${SCP} -r core@${VM_IP}:/tmp/kernel/* $INSTALL_DIR
-
-    ${SSH} core@${VM_IP} -- "sudo rm -fr /tmp/kernel"
-fi
-
-
 podman_version=$(${SSH} core@${VM_IP} -- 'rpm -q --qf %{version} podman')
 
 # Shutdown the VM
@@ -187,15 +160,6 @@ if [ "${SNC_GENERATE_LINUX_BUNDLE}" != "0" ]; then
     create_tarball "$libvirtDestDir"
 fi
 
-# vfkit image generation
-# This must be done after the generation of libvirt image as it reuses some of
-# the content of $libvirtDestDir
-if [ "${SNC_GENERATE_MACOS_BUNDLE}" != "0" ]; then
-    vfkitDestDir="${destDirPrefix}_vfkit_${destDirSuffix}"
-    rm -fr ${vfkitDestDir} ${vfkitDestDir}.crcbundle
-    generate_vfkit_bundle "$libvirtDestDir" "$vfkitDestDir" "$INSTALL_DIR" "$kernel_release" "$kernel_cmd_line"
-fi
-
 # HyperV image generation
 #
 # This must be done after the generation of libvirt image as it reuses some of
@@ -206,5 +170,43 @@ if [ "${SNC_GENERATE_WINDOWS_BUNDLE}" != "0" ]; then
     generate_hyperv_bundle "$libvirtDestDir" "$hypervDestDir"
 fi
 
-# Cleanup up vmlinux/initramfs files
-rm -fr "$INSTALL_DIR/vmlinuz*" "$INSTALL_DIR/initramfs*"
+# vfkit image generation
+# This must be done after the generation of libvirt image as it reuses some of
+# the content of $libvirtDestDir
+if [ "${SNC_GENERATE_MACOS_BUNDLE}" != "0" ]; then
+    start_vm ${VM_NAME} ${VM_IP}
+    if [ ${BUNDLE_TYPE} != "okd" ]; then
+        # workaround https://github.com/crc-org/vfkit/issues/11 on macOS 12
+        downgrade_rhel9_kernel ${VM_IP}
+        cleanup_vm_image ${VM_NAME} ${VM_IP}
+    fi
+
+    # Get the rhcos ostree Hash ID
+    ostree_hash=$(${SSH} core@${VM_IP} -- "cat /proc/cmdline | grep -oP \"(?<=${BASE_OS}-).*(?=/vmlinuz)\"")
+
+    # Get the rhcos kernel release
+    kernel_release=$(${SSH} core@${VM_IP} -- 'uname -r')
+
+    # Get the kernel command line arguments
+    kernel_cmd_line=$(${SSH} core@${VM_IP} -- 'cat /proc/cmdline')
+
+    # Get the vmlinux/initramfs to /tmp/kernel and change permission for initramfs
+    ${SSH} core@${VM_IP} -- "mkdir /tmp/kernel && sudo cp -r /boot/ostree/${BASE_OS}-${ostree_hash}/*${kernel_release}* /tmp/kernel && sudo chmod 644 /tmp/kernel/initramfs*"
+
+    # SCP the vmlinuz/initramfs from VM to Host in provided folder.
+    ${SCP} -r core@${VM_IP}:/tmp/kernel/* $INSTALL_DIR
+
+    ${SSH} core@${VM_IP} -- "sudo rm -fr /tmp/kernel"
+    shutdown_vm ${VM_NAME}
+
+    vfkitDestDir="${destDirPrefix}_vfkit_${destDirSuffix}"
+    rm -fr ${vfkitDestDir} ${vfkitDestDir}.crcbundle
+
+    create_bundle_qemu_image "$libvirtDestDir" "${VM_PREFIX}" "${VM_NAME}"
+
+    generate_vfkit_bundle "$libvirtDestDir" "$vfkitDestDir" "$INSTALL_DIR" "$kernel_release" "$kernel_cmd_line"
+
+    # Cleanup up vmlinux/initramfs files
+    rm -fr "$INSTALL_DIR/vmlinuz*" "$INSTALL_DIR/initramfs*"
+fi
+
